refactor: ContentsStorage 패키지 이동
